Force creation of session cookies before attempting authentication
For some reason the dns.he.net API does not seem to set a session cookie
when the very first request is already the login POST.

This commit adds a dummy GET request to the main URL to force creating a
session.  It then returns _that_ session cookies after the authentication
credentials have been POSTed.